+++ griffin/griffin.c
@@ -882,11 +882,11 @@ NETPLAY
 ============================================================ */
 #ifdef HAVE_NETWORKING
 #include "../network/netplay/netplay_frontend.c"
+#include "../network/netplay/netplay_init.c"
 #include "../network/netplay/netplay_io.c"
 #include "../network/netplay/netplay_net.c"
 #include "../network/netplay/netplay_common.c"
 #include "../network/netplay/netplay_discovery.c"
-#include "../network/netplay/netplay.c"
 #include "../libretro-common/net/net_compat.c"
 #include "../libretro-common/net/net_socket.c"
 #include "../libretro-common/net/net_http.c"
